Switches to left action, fixes order of operations bug
Switches to left action, so A*B is the funciton composition A o B.
(Previously, was B o A). Now the examples from Cannon-Floyd-Perry
can be typed in by the user directly from the paper.
Also fixes an order of operations bug introduced by previous commit,
where in A*B^3, * would be applied before ^.